Helpers class to POI-class
