+++ recommender-back/src/apis/poi.py
@@ -1,160 +1,59 @@
-from apis import weather
-from apis import helpers
-import json
-import requests
-import os
+from apis import times
 
 
-def merge_json(paths):
-    """
-    Merges json files together.
+class PointOfInterest:
+    def __init__(self, name=None, longitude=None, latitude=None, not_accessible_for=None, categories=None):
+        self.sun = times.get_sun_data()
+        self.name = name
+        self.longitude = longitude
+        self.latitude = latitude
+        self.not_accessible_for = not_accessible_for
+        self.categories = categories
+        self.weather = {}
 
-    Args:
-        paths: list of file paths
+    def calculate_score(self):
+        '''
+        Calculates a score for the POI based on weather data and time constraints..
+        '''
 
-    Returns:
-        List: json files merged together as a list.
-    """
-    merged = []
-    for path in paths:
-        with open(path, 'r') as json_file:
-            data = json.load(json_file)
-            merged = merged + data
-    return merged
+        sunrise = self.sun[0]
+        sunset = self.sun[1]
+        suitable_temperature_range = (20, 30)
+        suitable_humidity_range = (40, 60)
+        for timeinterval, data in enumerate(self.weather.values()):
+            time = times.get_current_time(timeinterval)
+            temperature_str = data.get('Air temperature')
+            humidity_str = data.get('Humidity')
+            try:
+                temperature = float(temperature_str.split('°C')[0])
+                humidity = float(humidity_str.split('%')[0])
+            except (TypeError, ValueError):
+                data['Score'] = -float('inf')
+                continue
+            except (AttributeError):
+                continue
+            if time >= sunrise and time <= sunset:
+                if suitable_temperature_range[0] <= temperature <= suitable_temperature_range[1]:
+                    temperature_score = 1.0
+                else:
+                    temperature_score = 0.0
 
+                if suitable_humidity_range[0] <= humidity <= suitable_humidity_range[1]:
+                    humidity_score = 1.0
+                else:
+                    humidity_score = 0.0
 
-def get_pois_as_json(accessibility=False, time=None):
-    """
-    Retrieves points of interest (POIs) from a JSON file and enriches them with current weather data.
+                score = (temperature_score + humidity_score) / 2
+                data['Score'] = score
+            else:
+                data['Score'] = 0.0
 
-    Returns:
-        str: JSON string containing the POIs with weather information.
+    def get_json(self):
+        '''
+        Returns a JSON representation of the POI.
 
-    Raises:
-        KeyError: If an error occurs while processing the data.
-
-    """
-    try:
-        pois = get_pois()
-        weather_data = weather.get_current_weather()
-        url = os.environ.get('REACT_APP_BACKEND_URL') + '/api/forecast'
-        response = requests.get(url)
-        forecast_data = response.json()
-        updated_data = []
-        for poi in pois:
-            poi = find_nearest_stations_weather_data(poi, weather_data)
-            poi = add_forecast_to_poi(poi, forecast_data)
-            if accessibility not in poi["accessibility_shortcoming_count"]:
-                updated_data.append(poi)
-            poi = helpers.Recommender(**poi)
-        return json.dumps(updated_data)
-    except KeyError as error:
-        return {
-            'message': 'An error occurred',
-            'status': 500,
-            'error': str(error)
-        }
-
-
-def find_nearest_stations_weather_data(poi, weather_data):
-    """
-    Finds the nearest weather station to a given point of interest (POI) and adds its weather data to the POI.
-
-    Args:
-        poi (dict): The POI for which weather data needs to be added.
-        weather_data (dict): A dictionary containing weather data for different weather stations.
-
-    Returns:
-        dict: The modified POI with weather information.
-
-    """
-    lat = float(poi['location']['coordinates'][1])
-    lon = float(poi['location']['coordinates'][0])
-    smallest, nearest = float('inf'), ''
-    for station in weather_data:
-        dist = abs(weather_data[station]['Longitude'] - lon)\
-            + abs(weather_data[station]['Latitude'] - lat)
-        if dist < smallest:
-            smallest, nearest = dist, station
-    poi['weather'] = {}
-    poi['weather']["Current"] = weather_data[nearest]
-    return poi
-
-
-def add_forecast_to_poi(poi, forecast_data):
-    """
-    Adds forecast data to the poi
-
-    Args:
-        poi (dict): The POI for which forecast data needs to be added.
-        forecast_data (dict): A dictionary containing forecast data for wanted coordinates.
-
-    Returns:
-        dict: The modified POI with forecast information.
-
-    """
-    lat = float(poi['location']['coordinates'][1])
-    lon = float(poi['location']['coordinates'][0])
-    for hour in forecast_data:
-        data = forecast_data[hour]
-        poi["weather"][f'{hour[11:16]}'] = data[f"{lat}, {lon}"]
-    return poi
-
-
-def get_closest_poi_coordinates_data(coordinates, data):
-    """
-    Finds the nearest coordinates forecast data for all of the POI's coordinates. Used for caching only
-    the nearest coordinates to the POI's.
-
-    Args:
-        coordinates (list): List of coordinates for the POI.
-        data (dict): A dictionary containing forecast data for different hours and coordinates.
-
-    Returns:
-        dict: A dictionary containing the nearest coordinates forecast data for each hour.
-
-    """
-    returned_data = {hour: {} for hour in data}
-    pois = get_pois()
-    closest_coordinates = {}
-    for poi in pois:
-        smallest = float('inf')
-        nearest = []
-        lat = float(poi['location']['coordinates'][1])
-        lon = float(poi['location']['coordinates'][0])
-        for coordinate in coordinates:
-            dist = abs(coordinate[0] - lat)\
-                + abs(coordinate[1] - lon)
-            if dist < smallest:
-                smallest = dist
-                nearest = [coordinate[0], coordinate[1]]
-        closest_coordinates[(
-            f"({nearest[0]}, {nearest[1]})")] = f"{lat}, {lon}"
-        for hour in data:
-            for key, value in closest_coordinates.items():
-                forecast = data[hour][key]
-                returned_data[hour][f"{value}"] = weather.parse_forecast(
-                    forecast)
-    return returned_data
-
-
-def get_pois(category=None):
-    """
-    Retrieves all points of interest (POIs) from JSON files and merges them together.
-
-    Args:
-        category (list): List of categories of POIs to retrieve. If None, default categories will be used.
-
-    Returns:
-        list: List of all POIs.
-
-    """
-    if category is None:
-        category = ['open_air_water', 'fitness_parks', 'athletics_venues', 'sports_halls']
-    paths = [
-        f"src/apis/poi_data/sports_and_physical/water_sports/{category[0]}.json",
-        f"src/apis/poi_data/sports_and_physical/outdoor_sports/neighborhood_sports/{category[1]}.json",
-        f"src/apis/poi_data/sports_and_physical/outdoor_sports/{category[2]}.json",
-        f"src/apis/poi_data/sports_and_physical/indoor_sports/{category[3]}.json"
-    ]
-    return merge_json(paths)
+        Returns:
+            dict: JSON representation of the POI.
+        '''
+        return {'name': self.name, 'weather': self.weather,
+                'latitude': self.latitude, 'longitude': self.longitude}
